+++ evennia/comms/models.py
@@ -22,7 +22,6 @@ necessary to easily be able to delete connections on the fly).
 from django.conf import settings
 from django.db import models
 from django.utils import timezone
-
 from evennia.comms import managers
 from evennia.locks.lockhandler import LockHandler
 from evennia.typeclasses.models import TypedObject
@@ -151,7 +150,7 @@ class Msg(SharedMemoryModel):
     db_header = models.TextField("header", null=True, blank=True)
     # the message body itself
     db_message = models.TextField("message")
-    # send date
+    # send date (note - this is in UTC.  Use the .date_created property to get it in local time)
     db_date_created = models.DateTimeField(
         "date sent", editable=False, auto_now_add=True, db_index=True
     )
@@ -193,6 +192,11 @@ class Msg(SharedMemoryModel):
     @lazy_property
     def tags(self):
         return TagHandler(self)
+
+    @property
+    def date_created(self):
+        """Return the field in localized time based on settings.TIME_ZONE."""
+        return timezone.localtime(self.db_date_created)
 
     # Wrapper properties to easily set database fields. These are
     # @property decorators that allows to access these fields using